+++ src/main/php/interceptor/PostInterceptor.php
@@ -20,8 +20,9 @@ interface PostInterceptor
     /**
      * does the postprocessing stuff
      *
-     * @param  \stubbles\webapp\Request   $request   current request
-     * @param  \stubbles\webapp\Response  $response  response to send
+     * @param   \stubbles\webapp\Request   $request   current request
+     * @param   \stubbles\webapp\Response  $response  response to send
+     * @return  bool
      */
-    public function postProcess(Request $request, Response $response);
+    public function postProcess(Request $request, Response $response): bool;
 }