+++ src/main/php/interceptor/PreInterceptor.php
@@ -20,8 +20,9 @@ interface PreInterceptor
     /**
      * does the preprocessing stuff
      *
-     * @param  \stubbles\webapp\Request   $request   current request
-     * @param  \stubbles\webapp\Response  $response  response to send
+     * @param   \stubbles\webapp\Request   $request   current request
+     * @param   \stubbles\webapp\Response  $response  response to send
+     * @return  bool
      */
-    public function preProcess(Request $request, Response $response);
+    public function preProcess(Request $request, Response $response): bool;
 }